fix: move snapshot code to dedicated file following project conventions
