+++ backend/api/src/api/main.py
@@ -11,7 +11,7 @@ from pydantic import BaseModel
 from .auth import auth
 from .config import FRONTEND_URL, TITLE, VERSION
 from .repos import deployed, staging
-from .snapshots import router as snapshots_router
+from .snapshots import snapshots
 
 
 async def _session_pruner():
@@ -54,7 +54,7 @@ class Health(BaseModel):
 app.include_router(auth.router)
 app.include_router(staging.router)
 app.include_router(deployed.router)
-app.include_router(snapshots_router)
+app.include_router(snapshots.router)
 
 
 @app.get("/", operation_id="rootInfo", response_model=RootInfo)
